fix logging naming bug when doing dataset split
